debug

+++ mmdet/core/bbox/bbox_target.py
@@ -47,7 +47,7 @@ def bbox_target_single(pos_bboxes,
     num_samples = num_pos + num_neg
     labels = pos_bboxes.new_zeros(num_samples, dtype=torch.long)
     label_weights = pos_bboxes.new_zeros(num_samples)
-    size_labels = pos_bboxes.new_zeros(num_samples, dtype=torch.float)
+    size_labels = pos_bboxes.new_zeros(num_samples, dtype=torch.long)
     size_labels_weights = pos_bboxes.new_zeros(num_samples)
     bbox_targets = pos_bboxes.new_zeros(num_samples, 4)
     bbox_weights = pos_bboxes.new_zeros(num_samples, 4)
@@ -63,7 +63,7 @@ def bbox_target_single(pos_bboxes,
         area = (pos_gt_bboxes[:,3] - pos_gt_bboxes[:,1])  * (pos_gt_bboxes[:,2]-pos_gt_bboxes[:,0])
         #print("area {}".format(area))
         gt_size = torch.where(area<64*64,torch.full_like(area,0),torch.full_like(area,1))
-        size_labels[:num_pos] = gt_size.float()
+        size_labels[:num_pos] = gt_size
         #print("gt size {}".format(gt_size))
         size_labels_weights[:num_pos] = pos_weight
     if num_neg > 0:
